Initial import from jeresig/trie-js.
Fix rogue endline that snuck through the build process.

Use Benchmark.js for performance testing.

Rewrite tests to use Benchmark.js. Add the while(true){} loops back in to make it easier to analyze the memory usage of the code.

Mention the blog post on JavaScript Trie Performance Analysis in the README.

Update to Benchmark.js 0.1.348.

Add in tests for succinct trie and binary search string.

Add a simple test for checking the lookup time of a Succinctly-stored Trie.

Change the ORD() function to an object for faster lookups, based upon a suggestion from Steve Hanov.

Fix finding repeated word bug

Finding 'aalsaals' returned true if 'aals' is in trie.